merge

+++ Algorithm.Python/Alphas/ShareClassMeanReversionAlpha.py
@@ -83,7 +83,7 @@ class ShareClassMeanReversionAlphaModel(AlphaModel):
 
     def __init__(self, *args, **kwargs):
         self.sma = SimpleMovingAverage(10)
-        self.position_window = RollingWindow[Decimal](2)
+        self.position_window = RollingWindow[float](2)
         self.alpha = None
         self.beta = None
         if 'tickers' not in kwargs:
